Time: 3 ms (36.42%), Space: 113.8 MB (27.98%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.cpp
@@ -5,19 +5,19 @@ public:
 
         if(n < k) return -1;
         //first window initialization from 0 to k-1;
-        int currentSum = 0;
+        double currentSum = 0;
         for(int i=0;i<k;i++){
             currentSum += nums[i];
         }
     
         //window is creating until we get our maxSum for maxAvg
-        int maxSum = currentSum;
+        double maxSum = currentSum;
         for(int right = k;right<n;right++){
-            int incoming = nums[right];
-            int outgoing = nums[right - k];
+            double incoming = nums[right];
+            double outgoing = nums[right - k];
             currentSum += incoming - outgoing;
             maxSum = std::max(maxSum,currentSum);
         }
-        return static_cast<double>(maxSum)/k;
+        return maxSum/k;
     }
 };
